Only generate benchmark input files once

+++ kernel_modules/benchmarks/tar_gz_decompress/generate.sh
@@ -1,4 +1,11 @@
-rm -f testfile testfile.tar.gz
+#!/bin/bash
+cd -- "$(dirname -- "${BASH_SOURCE:-$0}")"
+
+OUTPUT_FILE="testfile.tar.gz"
+
+if [ ! -f "$OUTPUT_FILE" ]; then
+echo "Generating $OUTPUT_FILE"
 base64 /dev/urandom | head -c 10000000 > testfile
-tar -czf testfile.tar.gz testfile
+tar -czf "$OUTPUT_FILE" testfile
 rm -f testfile
+fi
